remove dependency on old Hibernate validator
And, to this end, simplify AdditionalExceptionDetailsPrinter to use reflection to look for getter methods that can provide additional information.
All of the information provided by the old ExceptionDetailPrinters is still accessible this way, since in each case a getter method provided it.
In addition, this will provided information from other kinds of exceptions not known to this class beforehand.