upgrade Seam 2 dependency and remove old workarounds
These workarounds are not necessary in Seam 2.3.
(Also, the JbossClusterMonitorWorkaround doesn't compile because its superclass was removed in Seam 2.3)

Seam now looks for the UserTransaction in the JNDI location used by Jboss AS 7, so we don't need to use the JTS api directly anymore.

Also, mark the Seam dependency as optional. Many projects use util that don't use Seam.