remove the old parent pom
The abstract-java-project pom wasn't helping as much as I had originally thought.

* it locks down guava and joda-time versions (in a way that isn't impossible to override, but it does rule out setting these versions via an imported BOM)
* it adds a bunch of cruft that isn't applicable anymore (like clover settings and svn config)

 I want to phase out its usage from our projects.

It isn't too verbose hard to add the bits where it helped: guava, joda-time, hamcrest, and mockito dependencies, the java compiler version, and the project encoding settings.

This commit also upgrades guava (from r09 13.0.1) to and joda-time (from 1.6 to 2.0).
Since Iterables.reverse() was removed, this replaces its usage with Lists.reverse().
This also removes usage of a joda-time method that's now deprecated.

It also upgrade junit and mockito. The hamcrest dependency is altered from hamcrest-all to hamcrest-core and hamcrest-library.