straighten out build number plugin
The jar config used to come from a parent pom, but the previous commit removed it.
The jar is now built with an 'implementation version' of the first 5 digits of the git commit id.

There was also some wonkiness where the plugin was defined in <pluginManagement> and also in <plugins>, with different versions specified.