Update compiler target version to Java 8
The reason of the update is due to the issue noticed after eTimesheet migration. Class RequestEvent compiled for Java 1.6 had forced casting to Object[] with Lists.newArrayList() method in the constructor. The code was executed only with passing get parameters in the url so probably that is the reason why the issue was not noticed in CCP before. After packaging it with target version 1.8 the forced casting is not added by compiler and RequestEvent class works as expected. The version of the library was increased to 5 to do not override existing dependencies in other projects.