Remove exception recording/reporting functionality
This isn't needed by etimesheet or ccp anymore, as they both use red-egg/rollbar now.
Also, the RequestHistory didn't seem to serialize well in recent wildlfy versions;
I didn't dig into why.
(In one case there was a field that was oddly null, in another the serializer itself complained about something in the RequestHistory object that I don't recall.)

The memory-monitoring thing turned out to be not-very-useful, because it would often trigger a notice
when the heap was close to full, but only because GC hadn't run yet.
I think it'd be better to use datadog to monitor/alert on this.
It also depended on RequestHistory, which I wanted to remove.